Move sdc files to their own folder
