--- db/indexMdb.js
@@ -1,7 +1,0 @@
-const mongoose = require('mongoose');
-
-mongoose.connect('mongodb://localhost/videos');
-
-const db = mongoose.connection;
-
-module.exports = db;
