+++ db/sdc/indexMdb.js
@@ -1,0 +1,7 @@
+const mongoose = require('mongoose');
+
+mongoose.connect('mongodb://localhost/videodb');
+
+const db = mongoose.connection;
+
+module.exports = {db, mongoose};
